Organize constants
Moved boardState/goConstants into three files in the main Go folder:
Constants.ts: contains constants (and a couple functions for now)
Types.ts: contains types
Enums.ts: contains enums
Then updated imports to reflect new locations
