+++ src/Go/Constants.ts
@@ -1,33 +1,6 @@
-import { PartialRecord } from "../../Types/Record";
-import { getNewBoardState } from "./boardState";
-
-export enum GoColor {
-  white = "White",
-  black = "Black",
-  empty = "Empty",
-}
-
-export enum GoValidity {
-  pointBroken = "That node is offline; a piece cannot be placed there",
-  pointNotEmpty = "That node is already occupied by a piece",
-  boardRepeated = "It is illegal to repeat prior board states",
-  noSuicide = "It is illegal to cause your own pieces to be captured",
-  notYourTurn = "It is not your turn to play",
-  gameOver = "The game is over",
-  invalid = "Invalid move",
-  valid = "Valid move",
-}
-
-export enum GoOpponent {
-  none = "No AI",
-  Netburners = "Netburners",
-  SlumSnakes = "Slum Snakes",
-  TheBlackHand = "The Black Hand",
-  Tetrads = "Tetrads",
-  Daedalus = "Daedalus",
-  Illuminati = "Illuminati",
-  w0r1d_d43m0n = "????????????",
-}
+import { getNewBoardState } from "./boardState/boardState";
+import { GoOpponent } from "@enums";
+import type { OpponentStats, GoState } from "./Types";
 
 export const opponentDetails = {
   [GoOpponent.none]: {
@@ -96,101 +69,7 @@ export const opponentDetails = {
 
 export const boardSizes = [5, 7, 9, 13];
 
-export type Board = (PointState | null)[][];
-
-export type Move = {
-  point: PointState;
-  oldLibertyCount?: number | null;
-  newLibertyCount?: number | null;
-  createsLife?: boolean;
-};
-
-type MoveType =
-  | "capture"
-  | "defendCapture"
-  | "eyeMove"
-  | "eyeBlock"
-  | "pattern"
-  | "growth"
-  | "expansion"
-  | "jump"
-  | "defend"
-  | "surround"
-  | "corner"
-  | "random";
-
-type MoveFunction = () => Promise<Move | null>;
-export type MoveOptions = Record<MoveType, MoveFunction>;
-
-export type EyeMove = {
-  point: PointState;
-  createsLife: boolean;
-};
-
-export type BoardState = {
-  board: Board;
-  previousPlayer: GoColor | null;
-  history: Board[];
-  ai: GoOpponent;
-  passCount: number;
-  cheatCount: number;
-};
-
-export type PointState = {
-  player: GoColor;
-  chain: string;
-  liberties: (PointState | null)[] | null;
-  x: number;
-  y: number;
-};
-
-/**
- * "invalid" or "move" or "pass" or "gameOver"
- */
-export enum GoPlayType {
-  invalid = "invalid",
-  move = "move",
-  pass = "pass",
-  gameOver = "gameOver",
-}
-
-export type Play = {
-  success: boolean;
-  type: GoPlayType;
-  x: number;
-  y: number;
-};
-
-export type Neighbor = {
-  north: PointState | null;
-  east: PointState | null;
-  south: PointState | null;
-  west: PointState | null;
-};
-
-export type GoScore = {
-  White: { pieces: number; territory: number; komi: number; sum: number };
-  Black: { pieces: number; territory: number; komi: number; sum: number };
-};
-
-type OpponentStats = {
-  wins: number;
-  losses: number;
-  nodes: number;
-  nodePower: number;
-  winStreak: number;
-  oldWinStreak: number;
-  highestWinStreak: number;
-  favor: number;
-};
-
 export const columnIndexes = "ABCDEFGHJKLMNOPQRSTUVWXYZ";
-
-type GoState = {
-  previousGameFinalBoardState: BoardState | null;
-  boardState: BoardState;
-  status: PartialRecord<GoOpponent, OpponentStats>;
-};
 
 export function newOpponentStats(): OpponentStats {
   return {
