+++ src/Go/ui/GoGameboard.tsx
@@ -1,20 +1,22 @@
-import React, { useMemo } from "react";
-import Grid from "@mui/material/Grid";
+import type { BoardState } from "../Types";
 
+import React, { useMemo } from "react";
+import { Grid } from "@mui/material";
+
+import { GoOpponent, GoColor } from "@enums";
 import { getSizeClass, GoPoint } from "./GoPoint";
 import { useRerender } from "../../ui/React/hooks";
 import { boardStyles } from "../boardState/goStyles";
 import { getAllValidMoves, getControlledSpace } from "../boardAnalysis/boardAnalysis";
-import { BoardState, GoOpponent, GoColor } from "../boardState/goConstants";
 
-interface IProps {
+interface GoGameboardProps {
   boardState: BoardState;
   traditional: boolean;
   clickHandler: (x: number, y: number) => any;
   hover: boolean;
 }
 
-export function GoGameboard({ boardState, traditional, clickHandler, hover }: IProps): React.ReactElement {
+export function GoGameboard({ boardState, traditional, clickHandler, hover }: GoGameboardProps): React.ReactElement {
   useRerender(400);
 
   const currentPlayer =
@@ -35,37 +37,35 @@ export function GoGameboard({ boardState, traditional, clickHandler, hover }: IP
   const classes = boardStyles();
 
   return (
-    <>
-      <Grid container id="goGameboard" className={`${classes.board} ${traditional ? classes.traditional : ""}`}>
-        {boardState.board.map((row, y) => {
-          const yIndex = boardState.board[0].length - y - 1;
-          return (
-            <Grid container key={`column_${yIndex}`} item className={getSizeClass(boardSize, classes)}>
-              {row.map((point, x: number) => {
-                const xIndex = x;
-                return (
-                  <Grid
-                    item
-                    key={`point_${xIndex}_${yIndex}`}
-                    onClick={() => clickHandler(xIndex, yIndex)}
-                    className={getSizeClass(boardSize, classes)}
-                  >
-                    <GoPoint
-                      state={boardState}
-                      x={xIndex}
-                      y={yIndex}
-                      traditional={traditional}
-                      hover={hover}
-                      valid={pointIsValid(xIndex, yIndex)}
-                      emptyPointOwner={ownedEmptyNodes[xIndex]?.[yIndex]}
-                    />
-                  </Grid>
-                );
-              })}
-            </Grid>
-          );
-        })}
-      </Grid>
-    </>
+    <Grid container id="goGameboard" className={`${classes.board} ${traditional ? classes.traditional : ""}`}>
+      {boardState.board.map((row, y) => {
+        const yIndex = boardState.board[0].length - y - 1;
+        return (
+          <Grid container key={`column_${yIndex}`} item className={getSizeClass(boardSize, classes)}>
+            {row.map((point, x: number) => {
+              const xIndex = x;
+              return (
+                <Grid
+                  item
+                  key={`point_${xIndex}_${yIndex}`}
+                  onClick={() => clickHandler(xIndex, yIndex)}
+                  className={getSizeClass(boardSize, classes)}
+                >
+                  <GoPoint
+                    state={boardState}
+                    x={xIndex}
+                    y={yIndex}
+                    traditional={traditional}
+                    hover={hover}
+                    valid={pointIsValid(xIndex, yIndex)}
+                    emptyPointOwner={ownedEmptyNodes[xIndex]?.[yIndex]}
+                  />
+                </Grid>
+              );
+            })}
+          </Grid>
+        );
+      })}
+    </Grid>
   );
 }
